Refactors database handling into module
