+++ app/src/main/java/io/github/koss/mammut/feature/instance/InstanceActivity.kt
@@ -3,24 +3,18 @@ package io.github.koss.mammut.feature.instance
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.FragmentTransaction
 import com.bluelinelabs.conductor.Conductor
 import com.bluelinelabs.conductor.Router
 import com.bluelinelabs.conductor.RouterTransaction
 import io.github.koss.mammut.R
-import io.github.koss.mammut.data.repo.PreferencesRepository
+import io.github.koss.mammut.repo.PreferencesRepository
 import io.github.koss.mammut.extension.applicationComponent
 import io.github.koss.mammut.feature.base.BaseActivity
 import io.github.koss.mammut.feature.instance.dagger.InstanceComponent
 import io.github.koss.mammut.feature.instance.dagger.InstanceModule
-import io.github.koss.mammut.feature.instance.dagger.InstanceScope
 import io.github.koss.mammut.feature.instance.subfeature.navigation.InstanceController
-import io.github.koss.mammut.feature.instance.subfeature.navigation.ReselectListener
-import io.github.koss.mammut.feature.instance.subfeature.navigation.Tab
 import kotlinx.android.synthetic.main.activity_instance.*
-import org.jetbrains.anko.contentView
 import javax.inject.Inject
-import javax.inject.Named
 
 private const val EXTRA_INSTANCE_NAME = "instance_name"
 private const val EXTRA_AUTH_CODE = "auth_code"
@@ -84,7 +78,5 @@ class InstanceActivity : BaseActivity() {
                         context.startActivity(this)
                     }
         }
-
-        private const val STATE_CURRENT_TAB = "current_tab"
     }
 }
