+++ app/src/main/java/io/github/koss/mammut/feature/joininstance/JoinInstanceViewModel.kt
@@ -13,9 +13,9 @@ import com.sys1yagi.mastodon4j.api.method.Apps
 import com.sys1yagi.mastodon4j.api.method.Public
 import io.github.koss.mammut.R
 import io.github.koss.mammut.data.models.*
-import io.github.koss.mammut.data.repo.InstancesRepository
-import io.github.koss.mammut.data.repo.PreferencesRepository
-import io.github.koss.mammut.data.repo.RegistrationRepository
+import io.github.koss.mammut.repo.InstancesRepository
+import io.github.koss.mammut.repo.PreferencesRepository
+import io.github.koss.mammut.repo.RegistrationRepository
 import io.github.koss.mammut.extension.ClientBuilder
 import io.github.koss.mammut.extension.postSafely
 import io.github.koss.mammut.extension.run
