+++ app/src/main/java/io/github/koss/mammut/feature/themes/ThemeEngine.kt
@@ -5,7 +5,7 @@ import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.appbar.CollapsingToolbarLayout
 import io.github.koss.mammut.R
-import io.github.koss.mammut.data.repo.PreferencesRepository
+import io.github.koss.mammut.repo.PreferencesRepository
 import uk.co.chrisjenx.calligraphy.CalligraphyConfig
 
 class ThemeEngine(
