+++ data/README.md
@@ -1,0 +1,2 @@
+# Mammut Data Module
+This module is responsible for general data handling, i.e interacting with the in-app database. This has been extracted into it's own module for both encapsulation, as well as shared data access across modules.
