+++ data/src/main/java/io/github/koss/mammut/data/converters/EmojiConverter.kt
