+++ data/src/main/java/io/github/koss/mammut/data/database/dao/InstanceRegistrationDao.kt
