+++ data/src/main/java/io/github/koss/mammut/data/database/entities/InstanceRegistrationEntity.kt
