+++ data/src/main/java/io/github/koss/mammut/data/models/InstanceAccessToken.kt
