+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':instances', ':toot'
+include ':app', ':instances', ':toot', ':data'
